CameraScript.cs Minor Optiomization
The screen dimensions magnitude is calculated once in the Awake method and stored in screenDims. This will save computations in the FixedUpdate method, assuming that the screen size doesn't change frequently.

A displacement vector is created to be reused for each FixedUpdate to prevent frequent object creation and garbage collection. The components of displacement are updated directly in ApplyMouseDisplacement.

Removed the unnecessary Transform variable, transform1, in ApplyMouseDisplacement