WeaponRaycastLine.cs Huge Optimization
Used Object Pooling, this is especially applicable if you instantiate and destroy the WeaponRaycastLine frequently in your game. Create a pool of these objects and whenever you need one, you can get one from the pool, initialize it and use it. When you're done with it, you simply return it to the pool instead of destroying it.

Used Coroutines, coroutine creates a temporary overhead every time you call StartCoroutine. If you want to avoid this overhead, you could use a flag to check if the coroutine is running and update within FixedUpdate or Update instead of starting and stopping the coroutine.