Many Optimzations and List.RemoveAt Fix
In the Generate method, when picking a random index from the toBranch list, we are using List.RemoveAt which is O(n) as we noted in the comment. One solution to this is to swap the element to be removed with the last element of the list and then remove the last element, which is O(1).

Use Structs: Unity recommends using structs instead of classes when the data type can be a value type to reduce the garbage collection.

Use Readonly: If there's a variable that we know we're not going to change, it's a good idea to use the 'readonly' keyword. This will allow the compiler to make optimizations knowing that the value will not change.

Use Lists instead of arrays: Lists are easier to work with than arrays in C# and provide better performance in certain scenarios.

Precompute or cache the results of some calculations. This can be done if some calculations or methods are called repeatedly but always return the same value.

Decrease function calls: The Move() function is being called multiple times for the same values. This can be avoided by storing the result in a variable and using it instead of calling the function multiple times.

Reuse Random object: Currently, a new Random object is created each time the GenerateRNG() function is called. Instead, we can create the Random object once and reuse it, providing a new seed when necessary.